requirements.txt
asgiref              
Babel                
bitcoin             
certifi              
cffi                
charset-normalizer   
coverage             
crispy-bootstrap5    
cryptography         
Django               
django-crispy-forms       
django-extensions     
django-filter        
django-money         
django-widget-tweaks 
djangorestframework  
future               
idna                 
Markdown             
MarkupSafe           
pip                  
py-moneyed           
pycoin               
pycparser            
pyOpenSSL            
pytz                 
requests             
setuptools          
sqlparse             
transaction          
transactions         
typing_extensions    
tzdata               
urllib3             
Werkzeug             
wheel                
zope.interface       
